+++ packages/cli/src/cli.ts
@@ -1,25 +1,32 @@
-import { Command } from "commander";
 import { version } from "../package.json";
 import { build } from "./commands/build.js";
 import { init } from "./commands/init.js";
 import { watch } from "./commands/watch.js";
 
-const cli = new Command();
+const args = process.argv.slice(2);
+const command = args[0];
 
-cli.name("yummacss").description("Yumma CSS main command").version(version);
+switch (command) {
+	case "init":
+	case "i":
+		init();
+		break;
+	case "build":
+	case "b":
+		build().catch(() => process.exit(1));
+		break;
+	case "watch":
+	case "w":
+		watch().catch(() => process.exit(1));
+		break;
+	default:
+		console.log(`
+Yumma CSS v${version}
 
-cli.command("init").action(init).alias("i").description("Initialize config");
-
-cli
-	.command("build")
-	.action(() => build().catch(() => process.exit(1)))
-	.alias("b")
-	.description("Build styles once.");
-
-cli
-	.command("watch")
-	.action(() => watch().catch(() => process.exit(1)))
-	.alias("w")
-	.description("Build styles continuously.");
-
-cli.parse(process.argv);
+Commands:
+  init, i    Initialize config
+  build, b   Build styles once.
+  watch, w   Build styles continuously.
+`);
+		break;
+}