+++ packages/cli/tsdown.config.ts
@@ -3,7 +3,7 @@ import { defineConfig } from "tsdown";
 export default defineConfig({
 	banner: { js: "#!/usr/bin/env node" },
 	entry: ["src/index.ts", "src/cli.ts"],
-	format: ["esm", "cjs"],
+	format: ["esm"],
 	minify: true,
 	target: "es2020",
 });