+++ packages/core/src/defaults/variables.ts
@@ -1,7 +1,7 @@
 // borders
 export const borderWidth = 1;
-export const borderRadius = 0.25;
-export const borderSpacing = 0.25;
+export const borderRadius = .25;
+export const borderSpacing = .25;
 
 // outlines
 export const outlineOffset = 1;
@@ -33,19 +33,19 @@ export const transparent = "transparent";
 export const current = "currentColor";
 
 // box model
-export const height = 0.25;
-export const width = 0.25;
-export const padding = 0.25;
-export const margin = 0.25;
-export const spacing = 0.25;
+export const height = .25;
+export const width = .25;
+export const padding = .25;
+export const margin = .25;
+export const spacing = .25;
 
 // flexbox
-export const flexBasis = 0.25;
+export const flexBasis = .25;
 
 // grid
-export const gap = 0.25;
-export const columnGap = 0.25;
-export const rowGap = 0.25;
+export const gap = .25;
+export const columnGap = .25;
+export const rowGap = .25;
 
 // effect
 export const blur = 4;
@@ -72,10 +72,10 @@ export const boxShadowInset3xl = "inset 0 12px 24px 0 #00000033";
 
 export const grayscale = 10;
 
-export const opacity = 0.1;
+export const opacity = .1;
 
 // positioning
-export const topRightBottomLeft = 0.25;
+export const topRightBottomLeft = .25;
 
 // font
 export const fontMono =
@@ -85,8 +85,8 @@ export const fontSerif =
 export const fontDefault =
 	"system-ui, -apple-system, 'Segoe UI', Roboto, 'Helvetica Neue', 'Noto Sans', 'Liberation Sans', Arial, sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji', 'Segoe UI Symbol', 'Noto Color Emoji'";
 
-export const fontSizeXs = 0.75;
-export const fontSizeSm = 0.875;
+export const fontSizeXs = .75;
+export const fontSizeSm = .875;
 export const fontSizeMd = 1;
 export const fontSizeLg = 1.125;
 export const fontSizeXl = 1.25;