+++ packages/core/tsdown.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from "tsdown";
 
 export default defineConfig({
-	format: ["esm", "cjs"],
+	format: ["esm"],
 	minify: true,
 	target: "es2020",
 });