+++ packages/intellisense/tsdown.config.ts
@@ -3,7 +3,7 @@ import { defineConfig } from "tsdown";
 export default defineConfig({
 	dts: true,
 	entry: ["src/index.ts", "src/adapters/monaco.ts", "src/adapters/vscode.ts"],
-	format: ["esm", "cjs"],
+	format: ["esm"],
 	minify: true,
 	target: "es2020",
 	external: ["vscode"],